+++ runtime/src/pallets/mod.rs
@@ -6,3 +6,6 @@ pub use timestamp::*;
 
 pub mod authorship;
 pub use authorship::*;
+
+pub mod balances;
+pub use balances::*;
